worker-node/requirements.txt
requests
python-dotenv
scikit-learn
scipy
